Login class is updated
Password Encryption happens when the user attempts to create a new account and log in.
This class now checks the email format veridity at login and create account moment. 
If the email is invalid, error handling happens, in which the user cant get out of the retry loop until they enter the valid form email.