updated borrow and main to use it.
finalized borrow's checktime function to take current time and the time the user rented the book in order to find when the book is due back. Main now uses this to check if a user joins a waitlist for a book they have just rented now. still need to add to the join waitlist part themselves now tho.